# openapi.yaml
openapi: 3.0.0
info:
  title: Dokany API
  description: منصة دكاني - واجهات برمجة التطبيقات
  version: 1.1.0
  contact:
    name: Dokany Team
    url: https://dokany.com
servers:
  - url: https://api.dokany.com
    description: Production Server
  - url: http://localhost:3000
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    storeIdHeader:
      type: apiKey
      in: header
      name: x-store-id
  schemas:
...
    CartItem:
      type: object
      properties:
        productId:
          type: string
        variantSku:
          type: string
        quantity:
          type: integer
...
security:
  - bearerAuth: []
  - storeIdHeader: []
paths:
...
  /cart/sync:
    post:
      tags: [Cart]
      summary: مزامنة السلة
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: تمت المزامنة
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  syncedAt:
                    type: string
                  warnings:
                    type: array
                    items:
                      type: string
  /cart/validate:
    post:
      tags: [Cart]
      summary: التحقق من السلة
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/CartItem'
      responses:
        '200':
          description: نتيجة التحقق
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  validated:
                    type: array
                    items:
                      type: object
                      properties:
                        productId: { type: string }
                        valid: { type: boolean }
                        reason: { type: string }

  # ============================================================
  # 🎟️ COUPONS
  # ============================================================
...